Polish styling and add more calendar functionality
Component was styled to more closely resemble the original airbnb component. Additionally, the calendar has been given the functionality to display selected dates in a human-readable format, as well as store data for the dates in a format convenient for sending database queries.
The calendar can also detect the user trying to select an end date that is further back than the start date, but currently this does not work with dates outside the current month.